+++ src/onegov/translator_directory/views/translator.py
@@ -2,7 +2,6 @@ from datetime import datetime
 from io import BytesIO
 from onegov.file import File
 from morepath import redirect
-from docxtpl import DocxTemplate
 from morepath.request import Response
 from sedate import utcnow
 from onegov.core.custom import json
@@ -23,6 +22,7 @@ from onegov.translator_directory.forms.mail_templates import MailTemplatesForm
 from onegov.translator_directory.forms.mutation import TranslatorMutationForm
 from onegov.translator_directory.forms.translator import TranslatorForm, \
     TranslatorSearchForm, EditorTranslatorForm
+from onegov.translator_directory.generate_docx import fill_variables_in_docx
 from onegov.translator_directory.layout import AddTranslatorLayout,\
     TranslatorCollectionLayout, TranslatorLayout, EditTranslatorLayout,\
     ReportTranslatorChangesLayout, MailTemplatesLayout
@@ -458,50 +458,3 @@ def view_mail_templates(self, request, form):
         'title': _("Mail templates"),
         'button_text': _('Download')
     }
-
-
-def get_initials(first_name, last_name):
-    first_initials = first_name[:2].upper()
-    last_initials = last_name[:2].upper()
-    return last_initials + first_initials
-
-
-def fill_variables_in_docx(original_docx, t: Translator, **kwargs):
-    docx_template = DocxTemplate(original_docx)
-
-    # Variables to find and replace in final word file
-    substituted_variables = {
-        'translator_last_name': t.last_name,
-        'translator_first_name': t.first_name,
-        'translator_nationality': t.nationality,
-        'translator_gender': t.gender,
-        'translator_address': t.address,
-        'translator_city': t.city,
-        'translator_zip_code': t.zip_code,
-        'greeting': gendered_greeting(t),
-    }
-
-    # Values below are also required for one template. where to get?
-
-    # translator_decision = ('definitiv provisorisch).
-    # translator_function = ('Dolmetschen', 'Übsersetzen', '
-    # 'Kommunikationsüberwachung')
-
-    for key, value in kwargs.items():
-        substituted_variables[key] = value
-
-    docx_template.render(substituted_variables)
-    in_memory_docx = BytesIO()
-    docx_template.save(in_memory_docx)
-
-    in_memory_docx.seek(0)
-    return in_memory_docx.read()
-
-
-def gendered_greeting(translator):
-    if translator.gender == 'M':
-        return 'Sehr geehrter Herr'
-    elif translator.gender == 'F':
-        return 'Sehr geehrte Frau'
-    else:
-        return 'Sehr geehrte*r Herr/Frau'
